fix(member): 내부 API InterMemberController로 이동
