+++ src/main/java/com/bugzero/rarego/boundedContext/member/in/MemberController.java
@@ -1,6 +1,7 @@
 package com.bugzero.rarego.boundedContext.member.in;
 
 import com.bugzero.rarego.boundedContext.auction.app.AuctionFacade;
+
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -31,16 +32,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class MemberController {
 
-  private final MemberFacade memberFacade;
-  private final AuctionFacade auctionFacade;
-
-	@SecurityRequirement(name = "bearerAuth")
-	@Operation(summary = "소셜 로그인 이후 Member 생성(회원 가입)", description = "소셜 로그인 결과(email/provider)를 받아 회원가입 처리합니다.")
-	@PostMapping("/me")
-	public SuccessResponseDto<MemberJoinResponseDto> join(@RequestBody MemberJoinRequestDto requestDto) {
-		MemberJoinResponseDto responseDto = memberFacade.join(requestDto.email());
-		return SuccessResponseDto.from(SuccessType.CREATED, responseDto);
-	}
+	private final MemberFacade memberFacade;
 
 	@SecurityRequirement(name = "bearerAuth")
 	@Operation(summary = "회원 본인 조회", description = "본인의 정보를 조회합니다.")
